--- backend/alembic/versions/2025_04_29_0838-35af6c0314b9_add_server_name_and_updated_at_to_tools.py
@@ -1,58 +1,0 @@
-"""add_server_name_and_updated_at_to_tools
-
-Revision ID: 35af6c0314b9
-Revises: 7acc442922f2
-Create Date: 2025-04-29 08:38:05.768673
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision: str = '35af6c0314b9'
-down_revision: Union[str, None] = '7acc442922f2'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    # op.alter_column('sessions', 'user_id',
-    #            existing_type=mysql.BIGINT(),
-    #            nullable=False)
-    op.add_column('tools', sa.Column('server_name', sa.String(length=64), nullable=True))
-    op.add_column('tools', sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True))
-    # op.add_column('users', sa.Column('contacts', sa.JSON(), nullable=True))
-    # op.add_column('users', sa.Column('wallets', sa.JSON(), nullable=True))
-    # op.drop_index('email', table_name='users')
-    # op.drop_column('users', 'is_active')
-    # op.drop_column('users', 'email')
-    # op.drop_column('users', 'updated_at')
-    # op.drop_column('users', 'is_superuser')
-    # op.drop_column('users', 'full_name')
-    # op.drop_column('users', 'password_hash')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    # op.add_column('users', sa.Column('password_hash', mysql.VARCHAR(length=128), nullable=True))
-    # op.add_column('users', sa.Column('full_name', mysql.VARCHAR(length=120), nullable=True))
-    # op.add_column('users', sa.Column('is_superuser', mysql.TINYINT(display_width=1), autoincrement=False, nullable=True))
-    # op.add_column('users', sa.Column('updated_at', mysql.DATETIME(), nullable=True))
-    # op.add_column('users', sa.Column('email', mysql.VARCHAR(length=120), nullable=True))
-    # op.add_column('users', sa.Column('is_active', mysql.TINYINT(display_width=1), autoincrement=False, nullable=True))
-    # op.create_index('email', 'users', ['email'], unique=True)
-    # op.drop_column('users', 'wallets')
-    # op.drop_column('users', 'contacts')
-    op.drop_column('tools', 'updated_at')
-    op.drop_column('tools', 'server_name')
-    # op.alter_column('sessions', 'user_id',
-    #            existing_type=mysql.BIGINT(),
-    #            nullable=True)
-    # ### end Alembic commands ###
