--- backend/alembic/versions/2025_05_13_1751-41d1c68e9f9f_add_contacts_and_wallets_to_users.py
@@ -1,65 +1,0 @@
-"""add_contacts_and_wallets_to_users
-
-Revision ID: 41d1c68e9f9f
-Revises: 35af6c0314b9
-Create Date: 2025-05-13 17:51:34.911595
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision: str = '41d1c68e9f9f'
-down_revision: Union[str, None] = '35af6c0314b9'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    # 由于外键约束，不能修改user_id列的非空约束
-    # op.alter_column('sessions', 'user_id',
-    #            existing_type=mysql.BIGINT(),
-    #            nullable=False)
-    
-    # 添加新字段
-    op.add_column('users', sa.Column('contacts', sa.JSON(), nullable=True))
-    op.add_column('users', sa.Column('wallets', sa.JSON(), nullable=True))
-    
-    # password_hash可能已有数据，暂不修改nullable
-    # op.alter_column('users', 'password_hash',
-    #            existing_type=mysql.VARCHAR(length=128),
-    #            nullable=False)
-    
-    # 保留现有字段，以免影响已有数据
-    # op.drop_index('email', table_name='users')
-    # op.drop_column('users', 'is_active')
-    # op.drop_column('users', 'email')
-    # op.drop_column('users', 'full_name')
-    # op.drop_column('users', 'is_superuser')
-    # op.drop_column('users', 'updated_at')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    # op.add_column('users', sa.Column('updated_at', mysql.DATETIME(), nullable=True))
-    # op.add_column('users', sa.Column('is_superuser', mysql.TINYINT(display_width=1), autoincrement=False, nullable=True))
-    # op.add_column('users', sa.Column('full_name', mysql.VARCHAR(length=120), nullable=True))
-    # op.add_column('users', sa.Column('email', mysql.VARCHAR(length=120), nullable=True))
-    # op.add_column('users', sa.Column('is_active', mysql.TINYINT(display_width=1), autoincrement=False, nullable=True))
-    # op.create_index('email', 'users', ['email'], unique=True)
-    # op.alter_column('users', 'password_hash',
-    #            existing_type=mysql.VARCHAR(length=128),
-    #            nullable=True)
-    op.drop_column('users', 'wallets')
-    op.drop_column('users', 'contacts')
-    # op.alter_column('sessions', 'user_id',
-    #            existing_type=mysql.BIGINT(),
-    #            nullable=True)
-    # ### end Alembic commands ###
