--- backend/alembic/versions/2025_05_16_1828-7b9543a2c575_add_role_field_to_user_model.py
@@ -1,32 +1,0 @@
-"""Add role field to User model
-
-Revision ID: 7b9543a2c575
-Revises: 41d1c68e9f9f
-Create Date: 2025-05-16 18:28:27.586120
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision: str = '7b9543a2c575'
-down_revision: Union[str, None] = '41d1c68e9f9f'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('role', sa.Enum('user', 'developer', 'admin', name='user_role'), server_default='user', nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'role')
-    # ### end Alembic commands ###
