--- backend/alembic/versions/2025_08_29_1223-ecdcfce665b6_fix_foreign_key_type_mismatch.py
@@ -1,103 +1,0 @@
-"""fix_foreign_key_type_mismatch
-
-Revision ID: ecdcfce665b6
-Revises: bc7d9fe12ca3
-Create Date: 2025-08-29 12:23:34.310637
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision: str = 'ecdcfce665b6'
-down_revision: Union[str, None] = 'bc7d9fe12ca3'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('apps',
-    sa.Column('app_id', sa.String(length=64), nullable=False),
-    sa.Column('name', sa.String(length=128), nullable=False),
-    sa.Column('description', sa.String(length=512), nullable=True),
-    sa.Column('version', sa.String(length=32), nullable=False),
-    sa.Column('developer_id', sa.Integer(), nullable=False),
-    sa.Column('is_public', sa.Boolean(), nullable=False),
-    sa.Column('status', sa.Enum('draft', 'active', 'inactive', 'deployed', name='app_status'), nullable=False),
-    sa.Column('config', sa.JSON(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['developer_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('app_id')
-    )
-    op.create_table('app_tools',
-    sa.Column('id', sa.BigInteger(), autoincrement=True, nullable=False),
-    sa.Column('app_id', sa.String(length=50), nullable=False),
-    sa.Column('tool_id', sa.Integer(), nullable=False),
-    sa.Column('order_index', sa.Integer(), nullable=False),
-    sa.Column('config', sa.JSON(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['app_id'], ['apps.app_id'], ),
-    sa.ForeignKeyConstraint(['tool_id'], ['tools.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.alter_column('sessions', 'user_id',
-               existing_type=mysql.BIGINT(),
-               nullable=False)
-    op.add_column('tools', sa.Column('id', sa.Integer(), autoincrement=True, nullable=False))
-    op.add_column('tools', sa.Column('developer_id', sa.Integer(), nullable=True))
-    op.add_column('tools', sa.Column('is_public', sa.Boolean(), nullable=False))
-    op.add_column('tools', sa.Column('status', sa.Enum('active', 'inactive', 'pending', name='tool_status'), nullable=False))
-    op.add_column('tools', sa.Column('version', sa.String(length=32), nullable=False))
-    op.add_column('tools', sa.Column('tags', sa.JSON(), nullable=True))
-    op.add_column('tools', sa.Column('download_count', sa.Integer(), nullable=False))
-    op.add_column('tools', sa.Column('rating', sa.Float(), nullable=True))
-    op.create_foreign_key(None, 'tools', 'users', ['developer_id'], ['id'])
-    op.alter_column('users', 'is_active',
-               existing_type=mysql.TINYINT(display_width=1),
-               type_=sa.SmallInteger(),
-               existing_nullable=True)
-    op.alter_column('users', 'is_superuser',
-               existing_type=mysql.TINYINT(display_width=1),
-               type_=sa.SmallInteger(),
-               existing_nullable=True)
-    op.drop_index('email', table_name='users')
-    op.drop_column('users', 'contacts')
-    op.drop_column('users', 'wallets')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('wallets', mysql.JSON(), nullable=True))
-    op.add_column('users', sa.Column('contacts', mysql.JSON(), nullable=True))
-    op.create_index('email', 'users', ['email'], unique=True)
-    op.alter_column('users', 'is_superuser',
-               existing_type=sa.SmallInteger(),
-               type_=mysql.TINYINT(display_width=1),
-               existing_nullable=True)
-    op.alter_column('users', 'is_active',
-               existing_type=sa.SmallInteger(),
-               type_=mysql.TINYINT(display_width=1),
-               existing_nullable=True)
-    op.drop_constraint(None, 'tools', type_='foreignkey')
-    op.drop_column('tools', 'rating')
-    op.drop_column('tools', 'download_count')
-    op.drop_column('tools', 'tags')
-    op.drop_column('tools', 'version')
-    op.drop_column('tools', 'status')
-    op.drop_column('tools', 'is_public')
-    op.drop_column('tools', 'developer_id')
-    op.drop_column('tools', 'id')
-    op.alter_column('sessions', 'user_id',
-               existing_type=mysql.BIGINT(),
-               nullable=True)
-    op.drop_table('app_tools')
-    op.drop_table('apps')
-    # ### end Alembic commands ###
