+++ backend/alembic/versions/2025_09_02_1809-fd315aa4e6b2_initial_database_schema.py
@@ -1,0 +1,120 @@
+"""Initial database schema
+
+Revision ID: fd315aa4e6b2
+Revises: 
+Create Date: 2025-09-02 18:09:39.620955
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'fd315aa4e6b2'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('users',
+    sa.Column('id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('username', sa.String(length=64), nullable=False),
+    sa.Column('password_hash', sa.String(length=128), nullable=True),
+    sa.Column('email', sa.String(length=120), nullable=True),
+    sa.Column('full_name', sa.String(length=120), nullable=True),
+    sa.Column('role', sa.Enum('user', 'developer', 'admin', name='user_role'), nullable=False),
+    sa.Column('is_active', sa.SmallInteger(), nullable=True),
+    sa.Column('is_superuser', sa.SmallInteger(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
+    sa.Column('updated_at', sa.DateTime(), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('username')
+    )
+    op.create_table('apps',
+    sa.Column('app_id', sa.String(length=64), nullable=False),
+    sa.Column('name', sa.String(length=128), nullable=False),
+    sa.Column('description', sa.String(length=512), nullable=True),
+    sa.Column('version', sa.String(length=32), nullable=False),
+    sa.Column('developer_id', sa.BigInteger(), nullable=False),
+    sa.Column('is_public', sa.Boolean(), nullable=False),
+    sa.Column('status', sa.Enum('draft', 'active', 'inactive', 'deployed', name='app_status'), nullable=False),
+    sa.Column('config', sa.JSON(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
+    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
+    sa.ForeignKeyConstraint(['developer_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('app_id')
+    )
+    op.create_table('sessions',
+    sa.Column('session_id', sa.String(length=36), nullable=False),
+    sa.Column('user_id', sa.BigInteger(), nullable=False),
+    sa.Column('status', sa.Enum('interpreting', 'waiting_confirm', 'executing', 'done', 'error', name='session_status'), nullable=False),
+    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
+    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('session_id')
+    )
+    op.create_table('tools',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('tool_id', sa.String(length=64), nullable=False),
+    sa.Column('name', sa.String(length=128), nullable=False),
+    sa.Column('type', sa.Enum('mcp', 'http', name='tool_type'), nullable=False),
+    sa.Column('description', sa.String(length=512), nullable=True),
+    sa.Column('endpoint', sa.JSON(), nullable=False),
+    sa.Column('request_schema', sa.JSON(), nullable=False),
+    sa.Column('response_schema', sa.JSON(), nullable=True),
+    sa.Column('server_name', sa.String(length=64), nullable=True),
+    sa.Column('developer_id', sa.BigInteger(), nullable=True),
+    sa.Column('is_public', sa.Boolean(), nullable=False),
+    sa.Column('status', sa.Enum('active', 'inactive', 'pending', name='tool_status'), nullable=False),
+    sa.Column('version', sa.String(length=32), nullable=False),
+    sa.Column('tags', sa.JSON(), nullable=True),
+    sa.Column('download_count', sa.Integer(), nullable=False),
+    sa.Column('rating', sa.Float(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
+    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
+    sa.ForeignKeyConstraint(['developer_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('tool_id', 'server_name', name='uq_tools_tool_id_server')
+    )
+    op.create_table('app_tools',
+    sa.Column('id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('app_id', sa.String(length=50), nullable=False),
+    sa.Column('tool_id', sa.String(length=64), nullable=False),
+    sa.Column('order_index', sa.Integer(), nullable=False),
+    sa.Column('config', sa.JSON(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
+    sa.ForeignKeyConstraint(['app_id'], ['apps.app_id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_app_tools_app_id'), 'app_tools', ['app_id'], unique=False)
+    op.create_index(op.f('ix_app_tools_tool_id'), 'app_tools', ['tool_id'], unique=False)
+    op.create_table('logs',
+    sa.Column('id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('session_id', sa.String(length=36), nullable=False),
+    sa.Column('step', sa.String(length=32), nullable=False),
+    sa.Column('status', sa.String(length=32), nullable=False),
+    sa.Column('message', sa.Text(), nullable=True),
+    sa.Column('timestamp', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
+    sa.ForeignKeyConstraint(['session_id'], ['sessions.session_id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('logs')
+    op.drop_index(op.f('ix_app_tools_tool_id'), table_name='app_tools')
+    op.drop_index(op.f('ix_app_tools_app_id'), table_name='app_tools')
+    op.drop_table('app_tools')
+    op.drop_table('tools')
+    op.drop_table('sessions')
+    op.drop_table('apps')
+    op.drop_table('users')
+    # ### end Alembic commands ###
